Tratamento de exceção - delete

+++ src/main/java/com/empresa/curso/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.empresa.curso.entities.User;
 import com.empresa.curso.repositories.UserRepository;
+import com.empresa.curso.services.exceptions.DatabaseException;
 import com.empresa.curso.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,8 +32,15 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
-	}
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+        	repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+        	throw new DatabaseException("Violação de integridade de dados: " + e.getMessage());
+        }
+    }
 	
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
